+++ src/censo/modelos/ModeloImagen.java
@@ -10,11 +10,11 @@ package censo.modelos;
  * @author Admin
  */
 public class ModeloImagen {
-    private int id_imagen = 0;
+    private long id_imagen = 0;
     private String nombre = "";
     private byte[] imagen = null;
     
-    public void setIdImagen(int id_imagen){
+    public void setIdImagen(long id_imagen){
         this.id_imagen = id_imagen;
     }
     public void setNombre(String nombre){
@@ -24,7 +24,7 @@ public class ModeloImagen {
         this.imagen = imagen;
     }
     
-    public int getIdImagen(){
+    public long getIdImagen(){
         return id_imagen;
     }
     public String getNombre(){
